Feedback changes
As per the feedback received from Ritesh, changed the signature of the "view" function in "Article/Create.elm" module from "view: Model -> Html Msg" to "view: ApiKey -> Model -> Html Msg".
Needed to import "ApiKey" type from Admin.Request.Helper module.

Added an anchor tag in the view function to display the "Cancel" link with the href attribute set to the article listing page under the relative organization.

Removed the Cancel union type value and branch in the update function for the same.
Also, removed the "Cancel" button.

Provisioned the view function in the Main module to call the ArticleCreate.view function with an additional argument with the value of the organization's API key.

Removed the branch for ArticleCreate.Cancel in the update function from the Main module.