Widget "?" button should be displayed based on the settings
Added a "field" directive for the custom field "widget_visibility" in "app/graphql/types/organization_type.rb".
Registered its type as types.Boolean.
Preloaded associated setting record before resolving the value for this custom field.
Added a resolve method which fetches the resolved organization's setting record and returns the "enable" enum's value.

Also, added an additional "widgetVisibility" field to the Organization alias type in the "Client/Data/Organization.elm" module.
Provisioned the "organizationQuery" in the same module to build the GraphQL query by additionally requiring "widget_visibility" attribute of type bool which is then parsed as "widgetVisibility"  upon receiving the response.

Updated the "OrganizationLoaded (Ok organization)" branch in the update function of "Client/Main.elm" module to set the value of the "renderHelpButton" model field with the value of the "organization.widgetVisibility".