Cannot create an article if content is pasted in from clipboard
Investigated how to listen for the custom events on a DOM node.
Came across the "on" function in Html.Events module - https://github.com/elm/html/blob/bd01138f617459cc0299e9172e91dd7b48b5f7e9/src/Html/Events.elm#L200.
Studied an example about its usage.

Added an inline function "onTrixChange" in the view function of "Page.Article.Create" module.
Attached this function as the event handler on the node of the Trix editor instead of the "onInput" handler.
Imported Json.Decode module as Json which was required to decode the value received when the "trix-change" event gets triggered.

Elm compiler was unable to compile the changes due to the error: "This 'div' call produces: Html (String -> Msg). But all the previous elements in the list are: Html Msg".

Looked into how to keep the returned value of the 'div' elements consistent.