Custom event listener
Came across this comment on an issue which discussed about defining a custom event handler for the "change" event which matches the signature of the "inInput" built-in event handler - https://github.com/elm/html/issues/23#issuecomment-229192643.

Modified the inline "onTrixChange" function's signature and body to parse the event similar to the example posted in the above comment.

Tested the fix by pasting copied contents into the Trix editor and saving it.
Verified that the pasted contents gets saved and persisted in the database as expected.

Removed the "Debug.log" statement from the update function.