Present ticket listing in a tabular view
Modified the markup in the view function of Page.Ticket.List module from a DIV -> SPAN structure to the TABLE -> TD structure.
Added markup to display the table with appropriate Bootstrap 4.0 classes to so that rows appear highlighted when mouse is hovered on it.
Added markup for the heading section of table.
Added markup for the rows and provisioned it to display "View" link in the last column.