Add resize-aware fixed height to the editor
Removed the existing logic of pushing the editor upwards when the cursor reaches the bottom of the window because it was working in some cases and causing unexpected issues in some other cases.

Added "setTrixEditorHeight" function which calculates and sets the min-max height of the trix-editor node so as it can fit properly in the visible viewport.

Attached this function as an event listener on the events such as "trix-initialize" and "resize" of the window.

Added "overflow-y: scroll" CSS rule on the trix-editor node so that the overflowed content can be scrolled vertically using scrollbar.