Add the "attachments_path" in the articles GraphQL response
Added a custom field ":attachments_path" in the "app/graphql/types/article_type.rb".
Provisioned it to return the relative URL endpoint to upload the attachments for an article.

Added "attachmentsPath" property to the Article type alias in Admin.Data.Article module.
Modified the "articleObject" constant to parse the "attachments_path" from the response into the above "attachmentsPath" property.

Added "attachmentsPath" property to the Model type alias in Page.Article.Edit module as well.
Initialized that property with a blank string in initModel constant.
Updated the "ArticleLoaded (Ok articleResp)" branch in the "update" function to update the attachmentsPath property's value with the value received in the GraphQL response.
Added "data-attachments-path" attribute to the parent node of the "trix-editor" tag in the view function with the value of the "attachmentsPath" property in the model.

Updated the "uploadAttachement" function to accept an additional argument called "attachementsPath" and use it in the XHR request in the "app/javascript/packs/admin.js" script.
Also, added logic to retrieve the editor node's reference in the DOM in an event handler for the "trix-attachment-add" event.
Added logic to retrieve the value of the "data-attachments-path" attribute from the editor node's parent node.