Fix: Upload request fails with "Completed 401 Unauthorized"
Studied the #load_organization method in the "app/controllers/concerns/load_organization.rb" module.
Learned that we need to add the "api-key" header to the request while uploading an attachment.

Updated the signature of the view function in Page.Article.Edit to accept an additional argument of type ApiKey.
Added "data-api-key" attribute to the parent node of the "trix-editor" node in the view function with the value of this additional argument.
Updated the "update" function in the Admin.Main module to call the ArticleEdit.view function with an additional argument with the value of model.organizationKey.

Updated the event handler function of the "trix-attachment-add" event in the "app/javascript/packs/admin.js" script to fetch that "data-api-key" attribute's value and pass it to the "uploadAttachment" function while calling it.
Updated the "uploadAttachment" to set "api-key" request header of the XHR request with the value of this apiKey argument.

Verified that the upload request no longer fails with 401 status code.