Since upload request could fail on the backend, we need to ensure that user knows that the request was failed due to some reason.
Added logic in the "uploadAttachment" function to handle the response having status code other than 201 to be treated as an error response.
Provisioned to parse the "xhr.responseText" using JSON parser and try to fetch the "error" key from it and display it to the user.
If the responseText isn't a valid JSON, then provisioned it to display a generic error to the user.

Also, provisioned the "uploadAttachment" function to remove the attachment from the editor in case the upload request fails.