Extract out "uploadAttachment" function into a class
Added AttachmentUploader class.
Added a constructor function which accepts the same arguments which the previous "uploadAttachment" function was accepting.
Added "initializeRequest" function to it to initialize a new XHR request, set appropriate headers and attach necessary event handlers.
Added "formData" function to return a FormData object with the required formdata parameters.
Added "uploadProgressHandler" function to handle the attachment upload progress.
Added "uploadFinishHandler" function to handle the response received from the backend. For 201 status, provisioned it to call "uploadSuccessHandler" function, otherwise the "uploadErrorHandler" function for all all the other status codes.
Added "uploadSuccessHandler" function to handle the success response which retrieves the "attachment_url" from the response and sets the "url" and "href" attributes of the attachment object.
Added "uploadErrorHandler" function to handle error response, alert the user and remove the attachment from the editor.
Added the "upload" function to send the actual upload request to the backend with the form data.

Removed the "uploadAttachment" function from the "app/javascript/packs/admin.js" script.
Also, provisioned the event handler function for the "trix-attachment-add" event to issue an upload request using the new AttachmentUploader class instead of the old "uploadAttachment" function which no longer exists now.