Add a custom editor toolbar with an "attachments" button
Added "trixEditorToolbarView" function to the Article.Common module with the necessary markup to create a custom toolbar.
Provisioned the "trix-editor" nodes in both the Article.Create and Article.Edit modules to use this toolbar instead of the default toolbar.
Also, added a custom button to add attachments.
Provisioned it to trigger the provided Msg on clicking on that add attachments button.
Registered a Msg with value "AddAttachments" and added a branch to the update function to call the "addAttachments" outgoing port.
Registered the "addAttachments" port in the Ports module.
Subscribed to the "addAttachments" port in the "app/javascript/packs/admin.js" script and provisioned it to create a temporary file input tag and click on it to allow user to choose some files and then insert those files into the trix editor.
Also, added CSS to display the attachment icon on the attachment button in the Trix toolbar.