647 Add file upload support to the add article page
After the merge of https://github.com/bigbinary/acehelp/pull/648, rebase the https://github.com/bigbinary/acehelp/pull/647 with the latest master branch.

Added "attachmentsPath" field to the Model and initModel in Page.Article.Create module.

Updated the "ArticleLoaded (Ok articleResp)" branch in the update function to update the "attachmentsPath" field with the value of the loaded Article's response from the API.

Added "data-attachments-path" and "data-api-key" data-attributes to the parent node of the trie-editor node with the appropriate values.