Fix: Caption editing issue
As per the issue described in this video (https://github.com/bigbinary/acehelp/pull/647#issuecomment-425843343) related to editing of the uploaded image caption, the added caption is not displayed in the article's preview when article is saved.
Instead of the changed caption text, the article preview was displaying a textarea field along with a button with text "Remove" in it.

Tried to reproduce the issue.
After investigation, verified that this bug appears only while editing an existing article and not when a new article is saved.

Found that we are unnecessarily changing the the editor's description using the "trix-editor" DOM node's innertHTML property value in Elm using ports by listening for the "trix-change" event.
Removed the subscription for the trixChange port in the Page.Article.Edit module.
Removed the trixChange port in the Admin.Ports module.
Removed the event handler attached to the "trix-change" event in the "packs/admin.js" script which sends the editor's innerHTML property's value over the the trixChange port.