Calculate editor's height using Elm but set it using Ports
Directly setting the "style" attribute of the "trix-editor" node in the view function was actually setting those styles to the "trix-toolbar" node instead of the "trix-editor" node as expected.
Because of this weird issue, we decided to use Elm Ports to set the calculated style for the editor's height.

Have defined an outgoing port "setEditorHeight" in the Admin.Ports module.
Provisioned it to accept a record with fields such as articleId and height.
Subscribed to the "setEditorHeight" port on the JS side to find the editor element using the received "editorId" and set its min-height and max-height styles using the received "height"   option.

Removed "editorHeight" field from the model of Pages.Article.Edit module.
Also, removed the inline style attributes on the "trix-editor" node from the view function.

Modified the "ChangeEditorHeight (Ok info)" branch in the update function of the Pages.Article.Edit module to calculate the height using Browser.Dom package and instead of setting the "editorHeight" field in the model, provisioned it to call the "setEditorHeight" port.

Removed the old "setTrixEditorHeight" function which was pure JS function used to calculate the height and set the editor's height using that calculated value.