Fix: Editor height after clicking "Edit" button on an article show page
When an article show page is loaded, we set "isEditable" model field as false and therefore "display: none" style to the "trix-editor" node, therefore "trix-initialize" event doesn't get triggered.
After clicking the "Edit" button, the editor's height is still set to the default min-height inherited from the the article.scss stylesheet, which often is not what we want.

To fix the editor's height correctly in the above case, we now generate a delayed command which invokes the "setEditorHeight" outgoing port when user clicks the "Edit" button on the article edit page.