Adjust editor's height when window resizes
Added a subscription for the Browser.Events.onResize event in the both Pages.Article.Create and Pages.Article.Edit modules.
Defined ResizeWindow Msg.
Added a branch in the update function to handle this Msg.
Provisioned the ResizeWindow branch in the update function to call Dom.getElement and trigger the ChangeEditorHeight Cmd Msg after receiving the element's info.