Validate mime type and size of the file selected for upload
Added constants STATIC_IMAGE_MAX_SIZE_IN_BYTES, ANIMATED_IMAGE_MAX_SIZE_IN_BYTES and ALLOWED_ATTACHMENT_MIME_TYPES.

Since constant array is still mutable, frozenALLOWED_ATTACHMENT_MIME_TYPES array using Object.freeze function.

Added "validationError" constructor property to the AttachmentUploader class.

Added "validateAttachment" function.
Added a local "bytesToMBs" function in it.

Added a condition to validate that the attached file is allowed or not.

Added a validation to check the minimum size of the attached file.

Added a condition to check that the attached file is either of the types PNG or JPEG and has size less than 5 MB or not.

Added another validation check to ensure that the attached file is of type GIF and does not exceed 10 MB in size.

Updated the "upload" function to run the validations and check whether there was a validation error or not.
If there was a validation error, it removes the attachment and alerts the validation error message to the user; otherwise uploads the attachment to the backend.