Define PendingAction alias type
Defined and exposed PendingAction alias type in the Admin/Helpers.elm module.

Added "pendingActions" field of type "List PendingAction" to the Model in the Main module.

Modified the update function in the Main module to call the "ArticleCreate.update" function with an additional argument with the value of the "model.pendingActions" field.

Similarly, modified the view function in the Main module to call the "ArticleCreate.view" function with an additional argument with the value of the "model.pendingActions" field.

Modified the update and view functions in the Admin/Page/Article/Create.elm module accordingly to accept an additional argument of type "List PendingAction".