Add PendingActions opaque type
Added a new module PendingActions.

Defined PendingActions type constructor in it.
Also, defined Action type alias.

Implemented exists function which determines whether there exists an Action matching with the given id.

Implemented add function. Provisioned it to check if the specified id already exists or not.
If an Action record already exists, it returns the pending actions as it is.
Otherwise, it instantiates an Action record with the given id and message arguments and concatenates that action with the actions in the provided pendingActions argument.

Added remove function which returns the filtered PendingActions list by excluding the Action record in it if matches with the provided id argument.

Added "empty" constant which returns PendingActions with no an empty List of Action.

Implemented messages function which returns the list of strings containing values of each Action record's "message" property.

Implemented isEmpty function which determines whether the provided PendingActions argument contains any Action records or not.

Exposed all the functions and constants in the PendingActions module except the exists function.