Store PendingActions in the parent model and pass it down to sub-pages
Updated the Model in the Main module to store pendingActions field of type PendingActions.
Removed the previously defined PendingAction alias type from the Helpers module.
Updated the view and update functions in the Main module to pass down the pendingActions stored in the model to the view and update functions of the Pages.Article.Create module.
Updated the view and update functions in the Pages.Article.Create module accordingly.