Ability to add and remove pending actions for file uploads
Defined "addPendingAction" and "removePendingAction" incoming ports in the Admin.Ports module.
Provisioned the addPendingAction port to accept a record with fields such as id and message.

Provisioned the AttachmentUploader ES6 class to accept an additional constructor argument "finishCallback".
Defined "executeFinishCallback" function in it.
Also, extracted "removeAttachment" function to remove the attachment and execute the finishCallback.
Updated the "validateAttachment" function and "uploadErrorHandler" functions to call the "removeAttachment" function.
Also, updated the "uploadSuccessHandler" to call "executeFinishCallback" function when attachment is uploaded successfully.

Modified the "trix-attachment-add" event handler in the "javascript/packs/admin.js" script to define "pendingActionId" constant.
Also, defined "attachmentUploadFinishCallback" function which calls the "removePendingAction" port and passed this function as a constructor argument while initializing AttachmentUploader.
Before making the upload request, provisioned it to call the "addPendingAction" port with appropriate id and message.

Defined AddPendingAction and RemovePendingAction values for Msg type in the Main module.
Updated the subscriptions function to batch the page-specific subscriptions along with the subscriptions for the incoming ports such as "addPendingAction" and "removePendingAction".
Added branches for both AddPendingAction and RemovePendingAction Msg values in the update function to add a pending action with the specified id and message and remove the pending action matching the specified id respectively.