Add or remove a pending action when article contents is changed
Provisioned the branch for ArticleCreateMsg in the update "function" of Main module to pass-in the "model.pendingActions" to the "ArticleCreate.update" function.
Also, revised the logic for the same branch to update the "pendingActions" field of the model with the "pendingActions" field of the model of Page.Article.Create module.

Added "originalArticle" field to the model of Page.Article.Create module having the value of the type Maybe Article.
Updated the branches for SaveArticleResponse and ArticleLoaded messages in the update function to update the "originalArticle" field in the model with the article response received from the GraphQL API.

Defined "pendingActionsOnDescriptionChange" function in the Page.Article.Common module to update the provided pending actions by either adding a pending action when article contents is changed or remove the pending action when article contents is unchanged.
Also, defined "isDescriptionChanged" function to determine whether the new description is different from the original article's description before saving it.
Added "articleDescription" function which returns the actual description of the specified article argument.

Added "pendingActions" field to the model of Page.Article.Create module.
Revised the "DescInput" message branch in the update function of Page.Article.Create module to update the "pendingActions" field in the model using the "pendingActionsOnDescriptionChange" function.