Update the Page.Article.Edit to add or remove pending actions
Provisioned the ArticleEditMsg branch in the update function of the Main module to pass down the "model.pendingActions" to the model of the Page.Article.Edit module.
Also, updated the logic for the same branch to update the "pendingActions" field of the model with the "pendingActions" field of the model of Page.Article.Edit module.

Added "pendingActions" field to the model of Page.Article.Edit module.
Updated the "DescInput" Msg branch in the update function of Page.Article.Edit module to update the "pendingActions" field in the model using the "pendingActionsOnDescriptionChange" function.

Observed that when the edit article page is accessed in the non-editable form, the editor was still displayed on the page and was adding a pending action because it was getting mis-matching descriptions for the original article in the model and contents from the editor. To fix this case, added a conditional in the branch for "DescInput" Msg to check for the "model.isEditable" flag. If the "isEditable" flag is true, provisioned it to update the "pendingActions" otherwise, set "pendingActions" as an empty list.