Confirm user about pending actions while clicking on links
When there is at least a pending action and if user clicks on a link on the page, now we will present a confirmation dialog to the user and ask whether to ignore those pending actions or let them happen and do nothing.
If user chooses to ignore the pending actions, we will navigate the user to the link which user had clicked.

Defined "pendingActionsConfirmationDialog" view function in the Page.View module to render the confirmation dialog by showing the messages for each of the pending actions.

Added "showPendingActionsConfirmation" field to the model of Main module of type Acknowledgement.
Added IgnorePendingActions and HidePendingActionsConfirmationDialog messages.
Revised the branch for the "LinkClicked" Msg in the update function to check for the existence of pendingActions. If there are pendingActions, provisioned it to update the "showPendingActionsConfirmation" field in the model with "Yes (IgnorePendingActions (LinkClicked urlRequest))" value.

Handled the cases for the newly introduced messages such as IgnorePendingActions and HidePendingActionsConfirmationDialog respectively.

Updated the view function in the Main module to render the confirmation dialog using the "pendingActionsConfirmationDialog" function.