Prevent submitting article when uploads are pending
On new article page, upon clicking on the "Create Article" page, we will now display a confirmation dialog to the user if there are some pending actions.
If there exists at least one pending file upload action, we will display the confirmation dialog to the user asking whether to wait until all the remaining pending actions are full-filled or ignore the pending actions and save the article.
If there exists a pending action for the unsaved contents for the current article, we won't display the confirmation dialog to the user since user is trying to save that unsaved contents.

To achieve this, needed to make following changes.

Added "without" function in the PendingActions module which returns all the pending actions without the one which matches the specified pending action id.

Added "unsavedArticlePendingActionId" function in the Page.Article.Common module to return the pending id for the unsaved contents action for the current article.

Added "IgnorePendingActions" and "HidePendingActionsConfirmationDialog" msgs in the Page.Article.Create module and added branches for both in the update function accordingly.

Also, added "showPendingActionsConfirmation" field to the Model of Page.Article.Create module.

Modified the logic for the "SaveArticle" branch in the update function to show the confirmation dialog if there exists pending actions other than the pending action for the unsaved contents for the same article.

Besides, modified the success and error branches for the "SaveArticleResponse" msg to update the "pendingActions" field with "PendingActions.empty".