Prevent updating article when uploads are pending
Similar to the previous commit, this commit prevents saving the article contents on the edit article page when there are pending actions other than the pending action for the unsaved article contents.

Since the logic was duplicated, extracted out reusable "preventSaveForPendingActions" function in the Page.Article.Common module which decides whether to display the confirmation dialog to the user or not when user tries to save article form on both new article page and edit article  page respectively.

In the Page.Article.Edit module, added "showPendingActionsConfirmation" field to the Model.
Added "IgnorePendingActions" and "HidePendingActionsConfirmationDialog" msgs along with the cases for both the msgs in the update function.
Modified the branch for "SaveArticle" msg to check whether to display the confirmation dialog using  "preventSaveForPendingActions" function and update the "showPendingActionsConfirmation" field in the model accordingly.

Also, updated the branch for the "ResetArticle" msg to display the confirmation dialog when there exists at least a single pending action. The "ResetArticle" msg  is emitted after clicking on the "Cancel" button.

Added markup for the confirmation dialog to the view function.

Besides, updated the "save" function in the both Page.Article.Create and Page.Article.Edit modules to update "showPendingActionsConfirmation" field's value in the model to "No".