Prevent from signing out when there exists a pending action
The "Logout" text displayed on the top right corner on the global navbar is a button wrapped inside a form instead of the link.
As of now, clicking on the "Logout" button/text was directly signing out the user even if the user has unsaved contents in the editor on the new or edit article page or there are some ongoing file uploads.

To fix this behavior, updated the branch for the "SignOut" msg in the Main module to display a confirmation dialog to the user when there is at least a pending action and ask whether to ignore those pending actions and sign out anyway or stay on the page until those pending actions are finished.