Remove "pendingActions" from the sub-modules
As per the feedback, removed the "pendingActions" field from the model in both the Page.Article.Create and Page.Article.Edit modules.

Since we still need pendingActions in both of these modules, needed to pass the "pendingActions" argument from the Main module to the "update" and "view" functions of these submodules respectively.

At certain places, we need to modify the pendingActions from the "update" function in these sub-modules.  Therefore, provisioned the "update" functions in both of the Page.Article.Create and Page.Article.Edit sub-modules to return an updated "pendingActions" value in the returned tuple.
Needed to update the return value from each of the branch in the update function accordingly.