Delete temporary article when user clicks the cancel button in create article page
Changed the markup of the "Cancel" button from an anchor link to a button.

Defined "DiscardArticle" and "DeleteTemporaryArticleResponse" values for the Msg type.

Registered an onClick handler on the "Cancel" button to emit the "DiscardArticle" msg.

Added a branch for the "DiscardArticle" msg in the update function.
Provisioned it to check whether there exist any pending actions or not.
If there exists a pending action, provisioned it to update the "showPendingActionsConfirmation" field in the model with "Yes (IgnorePendingActions DiscardArticle)" so that a confirmation dialog is shown to the user first and if user clicks on the "Yes" button on that confirmation dialog, all the pending actions gets removed and the original "DiscardArticle" msg is handled at the end.
If there does not exist a pending action, provisioned the case for the "DiscardArticle" to return a ReaderCmd to request the deletion of the current article using the "requestDeleteArticle" function and emit the "DeleteTemporaryArticleResponse" msg to handle the response of the delete request.

Added a branch for for the "DeleteTemporaryArticleResponse" msg in the update function and provisioned it to do nothing.

Handled the "ArticleCreate.DeleteTemporaryArticleResponse" msg in the Main module's update function and provisioned it to navigate to article listing page using the "updateNavigation" function.