Fix: Not able to copy the widget's embed code in Firefox browser
Changed the container tag for the widget code from "textarea" to "pre".

Needed to add some CSS to wrap the line and words since the widget's code was appearing in a single line inside the "pre" tag.

Also, changed the background color, text color and padding of the widget's code block appropriately.