Set "widget_installed" to true when an initial request is issued by widget
Investigated about which request is issued by the widget when it renders on the host website.
Learned that the "request" is passed-down to the GraphQL functions in the "context" hash.
In the #call method of  the Resolvers::OrganizationsSearch function, inspected the value of the "context[:request].headers['X-Client']" for the requests issued by the widget.
Added WIDGET_CLIENT_REQUEST_HEADER constant.
Added logic to fetch the value of the "X-Client" request header from the context hash and update the "widget_installed" field with true value of the setting record associated with the organization matched with the passed-in API key.