Extract out the logic to identify whether the request is issued by client or not
Added #request_client_is_widget? method in the GraphqlController which checks if the "X-Client" request header has the value "widget" or not.
Assigned the value of this method to the :request_client_is_widget key in the #context hash.

Removed the similar logic from the #call method of Resolvers::OrganizationsSearch GraphQL function and instead made use of the  : request_client_is_widget key from the context hash to make the decision whether to update the "widget_installed" field in the Setting record or not.