Show whether widget is installed or not on the Settings page
Added "widget_installed" field in "app/graphql/types/setting_type.rb".

Updated the Setting alias type in Admin.Data.Setting module with the definition for the "widgetInstalled" field.
Updated the "settingObject" constant in the same module to parse the response of the setting object and fetch the boolean field "widget_installed" from the response into the "widgetInstalled" field of Setting alias type.

Added "installed" field in the Model of the Admin.Page.Settings.Widget module.

Updated the branches for the msgs such as "LoadSetting" and "SaveSettingResponse" to update the "installed" field of the model with the value of the "widgetInstalled" field in the loaded setting response.

Added a view function "widgetInstallationAcknowedgement" which returns the markup with appropriate copy saying whether the widget is installed or not.

Updated the "widgetSettingsView" view function accordingly.