No need of X-Client header since recently we added separate controllers
 Recently, we added separate controllers for handling requests issued by widget and dashboard respectively in https://github.com/bigbinary/acehelp/pull/694.
Because of this, now we need not to worry about tagging each request separately to differentiate it by injecting an "X-Client" header with the request.

Removed the "clientName" constant and the "X-Client" header from the "requestOptions" function in the Client.Request.Helpers module and in the Admin.Request.Helper module respectively.

Added Utils::Constants class and added REQUEST_SOURCE_WIDGET and REQUEST_SOURCE_DASHBOARD constants to it.

Provisioned the #context method in the GraphqlExecution::DashboardController to include "request_source" key-value pair with the value as "dashboard".

Similarly, added #context method to the GraphqlExecution::WidgetController by extending the hash returned by the #context method of the super class with the 'request_source" key-value pair with the value set as "widget".

Added #request_issued_by_widget? method in Resolvers::OrganizationsSearch to determine whether the request in the context hash is issued by the widget or not.
Also, added #set_widget_installed method in Resolvers::OrganizationsSearch to update the "widget_installed" attribute value to true of the associated Setting record of the organization.
Updated the #call method according.