Add AppUrlCarrier
Added AppUrlCarrier.
Added #app_url class method to it which returns the base app URL depending on what kind of the app is.
For staging and production apps, we get the base app URL with the "HEROKU_APP_URL" environment variable.
For review apps, we get the name of review app which we need to postfix with the "herokuapp.com" TLD.
Also, provisioned the #app_url method to see if "APP_URL" environment variable is available if none of the above environment variables are available and use it.
Added fallback logic to retrieve the base app URL from the request, otherwise return "http://localhost:3000" if no condition matches.

Removed the #app_url method from the ApplicationHelper in favor of the AppUrlCarrier::app_url method.