Add tests for AppUrlCarrier
Added a test to test the app URL when the "HEROKU_APP_URL" environment variable is present.

Added a test to test the app URL when the "HEROKU_APP_NAME" environment variable is present.

Added a test to test the app URL when the "APP_URL" environment variable is present.

Added a test to test the app URL when no above environment variables is present but the request argument is present.

Added a test to test the app URL when no above environment variables is present as well as request argument is not present.